pyproject.toml
[project]
name = "openmath-llm"
version = "0.1.0"
description = "OpenMath Language Model Mathematics Ontology Compiler"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "User", email = "user@example.com"}
]

dependencies = [
    # LLM Inference
    "vllm>=0.6.0",
    "transformers>=4.40.0",
    "torch>=2.1.0",
    "accelerate>=0.25.0",

    # Math & Symbolic Computation
    "sympy>=1.12",
    "antlr4-python3-runtime>=4.11",  # For SymPy LaTeX parsing

    # Data Processing
    "datasets>=2.16.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # Retrieval
    "rank-bm25>=0.2.2",  # BM25 sparse retrieval for hybrid mode
    "requests>=2.31.0",  # HTTP client for Ollama API

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # Utilities
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "ipykernel>=6.25.0",
    "jupyterlab>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openmath"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
